EIP1-3753 - Use "orphanRemoval = true" instead of separate Repository classes
