+++ src/test/kotlin/uk/gov/dluhc/printapi/config/IntegrationTest.kt
@@ -31,10 +31,8 @@ import software.amazon.awssdk.services.s3.S3Client
 import uk.gov.dluhc.printapi.client.BankHolidayDataClient
 import uk.gov.dluhc.printapi.config.SftpContainerConfiguration.Companion.PRINT_REQUEST_UPLOAD_PATH
 import uk.gov.dluhc.printapi.config.SftpContainerConfiguration.Companion.PRINT_RESPONSE_DOWNLOAD_PATH
-import uk.gov.dluhc.printapi.database.repository.AddressRepository
 import uk.gov.dluhc.printapi.database.repository.AnonymousElectorDocumentRepository
 import uk.gov.dluhc.printapi.database.repository.CertificateRepository
-import uk.gov.dluhc.printapi.database.repository.DeliveryRepository
 import uk.gov.dluhc.printapi.database.repository.TemporaryCertificateRepository
 import uk.gov.dluhc.printapi.jobs.BatchPrintRequestsJob
 import uk.gov.dluhc.printapi.jobs.FinalRetentionPeriodDataRemovalJob
@@ -136,16 +134,19 @@ internal abstract class IntegrationTest {
     protected lateinit var certificateRepository: CertificateRepository
 
     @Autowired
-    protected lateinit var deliveryRepository: DeliveryRepository
-
-    @Autowired
-    protected lateinit var addressRepository: AddressRepository
-
-    @Autowired
     protected lateinit var temporaryCertificateRepository: TemporaryCertificateRepository
 
     @Autowired
     protected lateinit var anonymousElectorDocumentRepository: AnonymousElectorDocumentRepository
+
+    @Autowired
+    protected lateinit var testDeliveryRepository: TestDeliveryRepository
+
+    @Autowired
+    protected lateinit var testAddressRepository: TestAddressRepository
+
+    @Autowired
+    protected lateinit var testPrintRequestRepository: TestPrintRequestRepository
 
     @BeforeEach
     fun clearLogAppender() {
